+++ src/config/buttons/ButtonBaseConfig.ts
@@ -1,11 +1,13 @@
 import {CardConfigButton} from "../../card/CardConfigButton";
 import {ButtonColors} from "./ButtonColors";
+import {ConfigCheckUtils} from "../ConfigCheckUtils";
 
 export class ButtonBaseConfig {
 
     constructor(
         public readonly buttonText: string,
         public readonly buttonColors: ButtonColors,
+        public readonly newRow: boolean,
         public active: boolean) {
     }
 
@@ -16,6 +18,7 @@ export class ButtonBaseConfig {
         return new ButtonBaseConfig(
             config.button_text,
             ButtonColors.fromConfig(config.colors, defaultColors),
+            ConfigCheckUtils.checkBooleanDefaultFalse(config.new_row),
             config.active);
     }
 }