+++ src/renderer/WindRoseRenderer.ts
@@ -1,10 +1,13 @@
 import {WindRoseData} from "./WindRoseData";
+import {Element} from "@svgdotjs/svg.js";
 
 export interface WindRoseRenderer {
 
     drawEmptyWindrose(): void;
 
     drawWindRose(windRoseData: WindRoseData, animate: boolean): void;
+
+    drawBackgroundImage(): Element | undefined;
 
     animateRemoveGraphs(): boolean;